refactor
se elimina duplicidad
